# Optimize indexer