Before the change:
```
BenchmarkIndexer-12        447849        2738 ns/op       242 B/op        4 allocs/op
PASS
ok   k8s.io/kubernetes/vendor/k8s.io/client-go/tools/cache 2.451s
```

After the change:
```
BenchmarkIndexer-12    	  906784	 1274 ns/op	  50 B/op	  3 allocs/op
PASS
ok   k8s.io/kubernetes/vendor/k8s.io/client-go/tools/cache	2.148s
```

```release-note
NONE
```

/kind cleanup
/priority important-soon
/sig api-machinery

/assign @jpbetz @caesarxuchao