@jberkus 2021-10-05
@wojtek-t if this is really a 100% improvement in performance, seems like it should be worthy of a release note, no?