@wojtek-t 2021-10-05
> @wojtek-t if this is really a 100% improvement in performance, seems like it should be worthy of a release note, no?

I don't agree. Release notes are for users and/or cluster admins. This change doesn't impact either of those in any visible way.